test(kafka): fold unit tests into index.spec, add autoCommit integration tests
- Move the 7 Kafka adapter unit tests from the standalone
  test/unit/kafka.spec.js into test/unit/index.spec.js, matching the
  repo's single-unit-file convention. Behavior unchanged: they remain
  broker-free (kafkajs consumer mocked) and run in every job like all
  unit tests.
- Add a Kafka-only integration block (gated by ADAPTER=Kafka, so it runs
  only in the Kafka CI job) verifying the autoCommit option end-to-end
  against a real broker: produce a 100-message backlog once, then confirm
  both autoCommit:false and autoCommit:true consume all 100 from the
  beginning under separate consumer groups.

Verified locally: full Kafka integration suite passes (21 tests).
